Extract RequestsBroker
